Store the state of items between scenes
Implement a game state manager to store states of items. In each scene, there will be a state saver and state loader, which will be triggered when leave&come from&to a scene. The next thing is to fix the bug where there will be more than one player when teleporting back to tutorial map